Added Feature07
Added Feature07: Payroll.
It iterates over all the employees and then checks if the employee comes under the category of payable today(that is if he is Hourly type of employee then he needs to be paid on Friday) and then the output is given in the fpormat: employee id and salary payable to him.
I am paying flat employees on the last working day of month and commission and extra hourly compensation is paid every friday. Union dues are deducted accordingly.